improve ptrf

+++ ft_nb_len_base.c
@@ -6,13 +6,13 @@
 /*   By: thifranc <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/12 16:10:29 by thifranc          #+#    #+#             */
-/*   Updated: 2016/03/31 11:37:18 by thifranc         ###   ########.fr       */
+/*   Updated: 2017/10/24 12:46:39 by thifranc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_nb_len_base(int n, int base)
+int		ft_nb_len_base(long long int n, int base)
 {
 	if (0 <= n && n < base)
 		return (1);
